pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Rafał Łazicki <rafal.lazicki99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.2"
requests = "^2.28.1"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
pytest-django = "^4.5.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
Pillow = "^9.2.0"
sorl-thumbnail = "^12.8.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
autoflake = "^1.4"
autopep8 = "^1.6.0"
black = "^22.6.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
mypy = "^0.971"
absolufy-imports = "^0.3.1"
pytest-django = "^4.5.2"

[tool.black]
line-length = 88
exclude = "migrations/*"

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["requests", "django", "rest_framework"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["manage.py", "venv", "migrations", "node_modules", ".env"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
allow_redefinition = true
exclude = ["manage.py$", "venv"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
